apps/payment-service/pyproject.toml
[project]
name = "payment-service"
version = "0.1.0"
description = "payment-service for transaction & api integration"
authors = [{ name = "Temirlan Ilik", email = "0penguin@mail.ru" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"patterns @ file:../../packages/patterns",
	"sqlalchemy (>=2)",
	"fastapi (>=0.116.1,<0.117.0)",
	"uvicorn (>=0.35.0,<0.36.0)",
	"asyncpg (>=0.29)",
	"aiosqlite (>=0.19)",
	"alembic (>=1.12)",
	"python-dotenv (>=1.1.1,<2.0.0)",
	"greenlet (>=3.2.4,<4.0.0)",
	"pydantic (>=2.11.7,<3.0.0)",
	"email-validator (>=2.3.0,<3.0.0)",
	"prometheus-client (>=0.22.1,<0.23.0)",
	"redis (>=6.4.0,<7.0.0)",
	"aiohttp (>=3.12.15,<4.0.0)",
	"python-logstash-async (>=4.0.2,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
patterns = { develop = true }
